pyproject.toml
[tool.poetry]
name = "Agri"
version = "0.1.0"
description = "Poetry File for Agricultrual Website"
authors = ["SamCrabtreree"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
python-decouple = "^3.6"
django-ckeditor = "^6.4.1"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
